added more examples, API juggling
