+++ cookbook/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-mogwai = { path = "../mogwai", version = "^0.4" }
+mogwai = { path = "../crates/mogwai", version = "^0.5" }
 
 [dependencies.web-sys]
 version = "^0.3"
