+++ cookbook/src/counter.md
@@ -1,0 +1,9 @@
+# A simple counter
+One of the simplest widgets is the click counter. It's always a great introduction
+to the workings of a new GUI library.
+
+## Example Code
+
+```rust, ignore
+{{#include ../../examples/counter/src/lib.rs}}
+```
