+++ examples/counter/.gitignore
@@ -1,0 +1,3 @@
+/target
+**/*.rs.bk
+/pkg
