+++ examples/counter/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8">
+        <title>list-of-gizmos</title>
+        <link rel="stylesheet" href="style.css">
+    </head>
+    <body>
+        <div id="app" ></div>
+        <script type="module">
+         import init, { main } from './pkg/counter.js';
+         window.addEventListener('load', async () => {
+             await init();
+             await main("app");
+         });
+        </script>
+    </body>
+</html>
